requirements.txt
streamlit
requests
pandas
cryptography
streamlit-authenticator>=0.3.0
PyYAML
bcrypt
openpyxl
XlsxWriter
gspread==5.12.0
gspread-dataframe==3.3.0
oauth2client
google-auth
google-auth-oauthlib
google-auth-httplib2
numpy
plotly
urllib3>=1.26.0
pydantic>=2.0.0
streamlit-aggrid>=0.3.3
aiohttp>=3.8.0